// src/assets/components/types.ts
// types.ts

export type RootDrawerParamList = {
  Menu: undefined;
  MinhaConta: undefined;
  DadosPessoais: undefined;
  LoginRedirect: undefined;
  Login: undefined;
  CadastroScreen: undefined;
  Pedidos: undefined;
  Endereços: undefined;
  Cartoes: undefined;
  DetalhesPedido: { pedido: any };
  Início: undefined;
  Buscar: undefined;
  Ofertas: undefined;
  Carrinho: undefined;
  Audio: { categoryId: string; categoryName: string };
  Cordas: { categoryId: string; categoryName: string };
  Percussao: { categoryId: string; categoryName: string };
  Acessorios: { categoryId: string; categoryName: string };
  Teclas: { categoryId: string; categoryName: string };
};